Move node to its own file
